+++ packages/fiori-mcp-server/src/constant.ts
@@ -2,3 +2,4 @@ export const LATEST_UI5_VERSION = '1.136.7';
 export const GENERATE_FIORI_UI_APP_ID = 'generate-fiori-ui-app';
 export const ADD_PAGE = 'add-page';
 export const DELETE_PAGE = 'delete-page';
+export const CREATE_CONTROLLER_EXTENSION_FUNCTIONALITY_ID = 'create-controller-extension';