+++ packages/fiori-mcp-server/src/tools/functionalities/controller-extension/index.ts
@@ -2,7 +2,7 @@ import { CustomExtensionType, DirName } from '@sap/ux-specification/dist/types/s
 import type { ControllerExtensionPageType } from '@sap/ux-specification/dist/types/src';
 import { SapuxFtfsFileIO } from '../../../page-editor-api';
 import type {
-    ExecuteFunctionalitiesInput,
+    ExecuteFunctionalityInput,
     ExecuteFunctionalityOutput,
     FunctionalityHandlers,
     GetFunctionalityDetailsInput,
@@ -11,6 +11,7 @@ import type {
 } from '../../../types';
 import { getDefaultExtensionFolder, resolveApplication } from '../../utils';
 import { join } from 'path';
+import { CREATE_CONTROLLER_EXTENSION_FUNCTIONALITY_ID } from '../../../constant';
 
 /**
  * Retrieves the parameter definitions for creating a controller extension.
@@ -42,7 +43,7 @@ function getParameters(): { pageType: Parameter; pageId: Parameter; controllerNa
 }
 
 export const CREATE_CONTROLLER_EXTENSION_FUNCTIONALITY: GetFunctionalityDetailsOutput = {
-    functionalityId: 'create-controller-extension',
+    functionalityId: CREATE_CONTROLLER_EXTENSION_FUNCTIONALITY_ID,
     name: 'Add new controller extension by creating javascript or typescript file and updates manifest.json with entry',
     description:
         'Add new controller extension by creating javascript or typescript file and updates manifest.json with entry. Controller extensions allow users to extensiate default behaviour with custom controllers code.',
@@ -85,7 +86,7 @@ async function getFunctionalityDetails(input: GetFunctionalityDetailsInput): Pro
  * @param input - The input parameters for executing the functionality.
  * @returns A promise that resolves to the execution output.
  */
-async function executeFunctionality(input: ExecuteFunctionalitiesInput): Promise<ExecuteFunctionalityOutput> {
+async function executeFunctionality(input: ExecuteFunctionalityInput): Promise<ExecuteFunctionalityOutput> {
     const { parameters, appPath } = input;
     const { pageId, controllerName, pageType } = parameters;
     const project = await resolveApplication(appPath);