+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.1.32
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.38.6
+
+## 0.1.31
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.38.5
+
+## 0.1.30
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.38.4
+
+## 0.1.29
+
+### Patch Changes
+
+-   Updated dependencies [080480a]
+    -   @sap-ux/fe-fpm-writer@0.38.3
+
 ## 0.1.28
 
 ### Patch Changes