+++ tests/integration/adaptation-editor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux-private/adaptation-editor-tests
 
+## 0.0.7
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.14
+
 ## 0.0.6
 
 ### Patch Changes